♻️ refactor: api 폴더 삭제, favorites의 useEffect 수정
